JWT Auth, Registration and Login OK

+++ java_backend/src/main/java/musinco/webapp/java_backend/repositories/UserRepository.java
@@ -3,7 +3,8 @@ package musinco.webapp.java_backend.repositories;
 import musinco.webapp.java_backend.models.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<Users, Long> {
+public interface UserRepository extends JpaRepository<Users, Integer> {
     Users findFirstByUsername(String username);
     Users findFirstByEmail(String email);
+
 }
